requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.3.1
matplotlib>=3.4.3
tqdm>=4.62.2
scikit-learn>=0.24.2
pandas>=1.3.2
seaborn>=0.11.2
torchviz>=0.0.2
graphviz>=0.17
PyQt5>=5.15.4 